Module: Rename `Install` to `Add`
