+++ src/AbpHelper/AbpHelper.csproj
@@ -9,13 +9,13 @@
     <Copyright>EasyAbp Team</Copyright>
     <PackageProjectUrl>https://github.com/EasyAbp/AbpHelper.CLI</PackageProjectUrl>
     <RepositoryUrl>https://github.com/EasyAbp/AbpHelper.CLI</RepositoryUrl>
-    <PackageVersion>0.6.1</PackageVersion>
+    <PackageVersion>0.7.0</PackageVersion>
     <PackAsTool>true</PackAsTool>
     <ToolCommandName>abphelper</ToolCommandName>
     <RootNamespace>EasyAbp.AbpHelper</RootNamespace>
     <AssemblyName>EasyAbp.AbpHelper</AssemblyName>
-    <AssemblyVersion>0.6.1</AssemblyVersion>
-    <FileVersion>0.6.1</FileVersion>
+    <AssemblyVersion>0.7.0</AssemblyVersion>
+    <FileVersion>0.7.0</FileVersion>
     <Authors>EasyAbp Team</Authors>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <PackageTags>abp helper generator</PackageTags>
@@ -51,6 +51,6 @@
   
   <PropertyGroup>
     <SatelliteResourceLanguages>en</SatelliteResourceLanguages>
-    <Version>0.6.1</Version>
+    <Version>0.7.0</Version>
   </PropertyGroup>
 </Project>
