+++ src/AbpHelper/Commands/Module/Add/AddCommandOption.cs
@@ -1,0 +1,10 @@
+﻿using EasyAbp.AbpHelper.Attributes;
+
+namespace EasyAbp.AbpHelper.Commands.Module.Add
+{
+    public class AddCommandOption : ModuleCommandOption
+    {
+        [Option('v', "version", Description = "Specify the version of the package(s) to add")]
+        public string Version { get; set; } = null!;
+    }
+}
